requirements.txt
torch
torchvision
tensorflow
numpy
pillow
pyyaml==5.4.1